requirements_dev.txt
# Basic Packages:
pandas
numpy
matplotlib
seaborn
scikit-learn
ensure==1.0.2
dill
PyYAML
ipykernel

# Project Packages:
google-generativeai
langchain
langchain_google_genai
PyPDF2
faiss-cpu
chromadb
python-dotenv
streamlit

# Initiating Setup.py:
-e .

# Additional Modules:
pytest==7.1.3 # Python testing framework
tox==3.25.1  # Command-line driven automated testing tool
black==22.8.0 # Format the document
flake8==5.0.4 # Linter focuses on code style and quality
mypy==0.971 # Type checker that checks for type-related issues
